User/salma salhi/development (#6)
* Add files via upload

This is the functionality for reading from the file and creating a binary search tree. Make sure you've got the file input.txt in your project folder before you run the program.

* Add files via upload

This is the code for the client server stuff. Not sure if it works because it needs to be tested with Zach's GUI code, but it includes all the necessary functionality. Make sure that you also download Zach's student class on the Client end; I haven't implemented it here.

Co-authored-by: SalmaSalhi7 <49101315+SalmaSalhi7@users.noreply.github.com>